Move Get kanban to task controller and make it private
