+++ server/claim-review/claim-review.module.ts
@@ -7,7 +7,6 @@ import { UtilService } from "../util";
 import { SourceModule } from "../source/source.module";
 import { ConfigModule } from "@nestjs/config";
 import { HistoryModule } from "../history/history.module";
-import { ViewModule } from "../view/view.module";
 
 export const ClaimReviewModel = MongooseModule.forFeature([
     {
@@ -22,10 +21,9 @@ export const ClaimReviewModel = MongooseModule.forFeature([
         HistoryModule,
         SourceModule,
         ConfigModule,
-        ViewModule
     ],
     providers: [UtilService, ClaimReviewService],
     exports: [ClaimReviewService],
     controllers: [ClaimReviewController],
 })
-export class ClaimReviewModule {}
+export class ClaimReviewModule { }
